Fix non-functional MTP.
This requires a minor patch to the kernel:

  https://github.com/ianmacd/a51xx/commit/4a679c3be0fed34265d317107f7534fb29322438

and a kernel config change to avoid a reproducible spontaneous reboot
after 220 seconds of uptime in TWRP:

  https://github.com/ianmacd/a51xx/commit/45714f8885687a6cbc0292c067368c839e8b5d8c

The default TWRP init.rc is now used in favour of Samsung's recovery
init.rc, and a custom USB configuration is applied.

Additionally, MTP no longer needs to be disabled to allow adb to work.